Module->DAQModule and split the code
