+++ python/appfwk/conf_utils.py
@@ -35,6 +35,8 @@ import dunedaq.rcif.cmd as rccmd  # AddressedCmd,
 import dunedaq.trigger.moduleleveltrigger as mlt
 import dunedaq.networkmanager.nwmgr as nwmgr
 
+from appfwk.daqmodule import DAQModule
+
 console = Console()
 
 ########################################################################
@@ -52,25 +54,6 @@ console = Console()
 
 # TODO: Make these all dataclasses
 
-class Module:
-    """An individual DAQModule within an application, along with its
-       configuration object and list of outgoing connections to other
-       modules
-    """
-
-    def __init__(self, plugin, conf=None, connections=None, name="__module"):
-        self.plugin=plugin
-        self.name=name
-        self.conf=conf
-        self.connections=connections if connections else dict()
-
-    def __repr__(self):
-        return f"{self.name} module(plugin={self.plugin}, conf={self.conf}, connections={self.connections})"
-
-    def __rich_repr__(self):
-        yield "plugin", self.plugin
-        yield "conf", self.conf
-        yield "connections", self.connections
 
 
 Connection = namedtuple("Connection", ['to', 'queue_kind', 'queue_capacity', 'queue_name', 'toposort'], defaults=("FollyMPMCQueue", 1000, None, True))
@@ -102,248 +85,11 @@ class Endpoint:
 GeoID = namedtuple('GeoID', ['system', 'region', 'element'])
 FragmentProducer = namedtuple('FragmentProducer', ['geoid', 'requests_in', 'fragments_out', 'queue_name'])
 
-class ModuleGraph:
-    """A set of modules and connections between them.
-
-    modulegraph holds a dictionary of modules, with each module
-    knowing its (outgoing) connections to other modules in the
-    modulegraph.
-
-    Connections to other modulegraphs are represented by
-    `endpoints`. The endpoint's `external_name` is the "public" name
-    of the connection, which other modulegraphs should use. The
-    endpoint's `internal_name` specifies the particular module and
-    sink/source name which the endpoint is connected to, which may be
-    changed without affecting other applications.
-
-    """
-
-    def __init__(self, modules:[Module]=None, endpoints=None, fragment_producers=None):
-        self.modules=modules if modules else []
-        self.endpoints=endpoints if endpoints else dict()
-        self.fragment_producers = fragment_producers if  fragment_producers else dict()
-
-    def __repr__(self):
-        return f"modulegraph(modules={self.modules}, endpoints={self.endpoints}, fragment_producers={self.fragment_producers})"
-
-    def __rich_repr__(self):
-        yield "modules", self.modules
-        yield "endpoints", self.endpoints
-        yield "fragment_producers", self.fragment_producers
-
-    def set_from_dict(self, module_dict):
-        self.modules=module_dict
-
-    def digraph(self):
-        deps = nx.DiGraph()
-        modules_set = set()
-
-        for module in self.modules:
-            if module.name in modules_set:
-                raise RuntimeError(f"Module {module.name} appears twice in the ModuleGraph")
-            deps.add_node(module.name)
-            modules_set.add(module.name)
-
-        for module in self.modules:
-            from_module = module.name
-            for connection in module.connections.values():
-                conn_data = connection.to.split(".")
-                if len(conn_data) != 2:
-                    raise RuntimeError(f'Bad connection: {conn_data} must be specified as module.queue_name')
-                to_module = conn_data[0]
-                if to_module == from_module:
-                    raise RuntimeError(f'Bad connection: {conn_data} you are connecting a {from_module} to itself!')
-                queue_name = conn_data[1]
-                if to_module in modules_set:
-                    deps.add_edge(from_module, to_module, label=queue_name)
-                else:
-                    raise RuntimeError(f"Bad connection {connection}: internal connection which doesn't connect to any module! Available modules: {modules_set}")
-
-        # now moving on to external links
-        for endpoint in self.endpoints.values():
-            endpoint_internal_data = endpoint.internal_name.split(".")
-            if len(endpoint_internal_data) != 2:
-                raise RuntimeError(f'Bad endpoint!: {endpoint} internal_endpoint must be specified as module.queue_name')
-            to_module = endpoint_internal_data[0]
-            if to_module in modules_set:
-                if endpoint.direction == Direction.IN:
-                    deps.add_edge(endpoint.external_name, to_module, label=endpoint_internal_data[1])
-                else:
-                    deps.add_edge(to_module, endpoint.external_name, label=endpoint_internal_data[1])
-                deps.nodes[endpoint.external_name]['color'] = 'red'
-            else:
-                raise RuntimeError(f"Bad endpoint {endpoint}: internal connection which doesn't connect to any module! Available modules: {modules_set}")
-
-        # finally the fragment producers
-        for producer in self.fragment_producers.values():
-            producer_request_in = producer.requests_in.split(".")
-            if len(producer_request_in) != 2:
-                raise RuntimeError(f"Bad fragment producer {producer}: request_in must be specified as module.queue_name")
-
-            if producer_request_in[0] in modules_set:
-                deps.add_edge("TriggerRecordBuilder", producer_request_in[0], label='requests')
-                deps.nodes["TriggerRecordBuilder"]['color'] = 'red'
-            else:
-                raise RuntimeError(f"Bad FragmentProducer {producer}: request_in doesn't connect to any module! Available modules: {modules_set}")
-
-            producer_frag_out = producer.fragments_out.split(".")
-            if len(producer_frag_out) != 2:
-                raise RuntimeError(f"Bad fragment producer {producer}: fragments_out must be specified as module.queue_name")
-
-            if producer_frag_out[0] in modules_set:
-                deps.add_edge(producer_frag_out[0], "FragmentReceiver", label='fragments')
-                deps.nodes["FragmentReceiver"]['color'] = 'orange'
-            else:
-                raise RuntimeError(f"Bad FragmentProducer {producer}: fragments_out doesn't connect to any module! Available modules: {modules_set}")
-
-        return deps
-
-    def get_module(self, name):
-        for mod in self.modules:
-            if mod.name == name:
-                return mod
-        return None
-
-    def reset_module(self, name, new_module):
-        for mod in self.modules:
-            if mod.name == name:
-                mod = new_module
-                return
-        raise RuntimeError(f'Module {name} not found!')
-
-    def module_names(self):
-        return [n.name for n in self.modules]
-
-    def module_list(self):
-        return self.modules
-
-    def add_module(self, name, **kwargs):
-        if get_module(name):
-            raise RuntimeError(f"Module of name {name} already exists in this modulegraph")
-        mod=module(**kwargs)
-        self.modules.append(mod)
-        return mod
-
-    def add_connection(self, from_endpoint, to_endpoint):
-        from_mod, from_name=from_endpoint.split(".")
-        self.get_module(from_mod).connections[from_name]=Connection(to_endpoint)
-
-    def add_endpoint(self, external_name, internal_name, inout, topic=[]):
-        self.endpoints[external_name] = Endpoint(external_name, internal_name, inout, topic)
-
-    def endpoint_names(self, inout=None):
-        if inout is not None:
-            return [ e[0] for e in self.endpoints.items() if e[1].inout==inout ]
-        return self.endpoints.keys()
-
-    def add_fragment_producer(self, system, region, element, requests_in, fragments_out):
-        geoid = GeoID(system, region, element)
-        if geoid in self.fragment_producers:
-            raise ValueError(f"There is already a fragment producer for GeoID {geoid}")
-        # Can't set queue_name here, because the queue names need to be unique system-wide,
-        # but we're inside a particular app here. Instead, we create the queue names in readout_gen.generate,
-        # where all of the fragment producers are known
-        queue_name = None
-        self.fragment_producers[geoid] = FragmentProducer(geoid, requests_in, fragments_out, queue_name)
-
-class App:
-    """
-    A single daq_application in a system, consisting of a modulegraph
-    and a hostname on which to run
-    """
-
-    def __init__(self, modulegraph=None, host="localhost", name="__app"):
-        if modulegraph:
-            # hopefully that crashes if something is wrong!
-            self.digraph = modulegraph.digraph()
-            self.digraph.name = name
-
-        self.modulegraph = modulegraph if modulegraph else ModuleGraph()
-        self.host = host
-        self.name = name
-
-    def reset_graph(self):
-        if self.modulegraph:
-            self.digraph = self.modulegraph.digraph()
-
-    def export(self, filename):
-        if not self.digraph:
-            raise RuntimeError("Cannot export a app which doesn't have a valid digraph")
-        nx.drawing.nx_pydot.write_dot(self.digraph, filename)
 
 Publisher = namedtuple(
     "Publisher", ['msg_type', 'msg_module_name', 'subscribers'])
 
 Sender = namedtuple("Sender", ['msg_type', 'msg_module_name', 'receiver'])
-
-class System:
-    """
-    A full DAQ system consisting of multiple applications and the
-    connections between them. The `apps` member is a dictionary from
-    application name to app object, and the app_connections member is
-    a dictionary from upstream endpoint to publisher or sender object
-    representing the downstream endpoint(s). Endpoints are specified
-    as strings like app_name.endpoint_name.
-
-    An explicit mapping from upstream endpoint name to zeromq
-    connection string may be specified, but typical usage is to not
-    specify this, and leave the mapping to be automatically generated.
-
-    The same is true for application start order.
-    """
-
-    def __init__(self, apps=None, app_connections=None, network_endpoints=None, app_start_order=None):
-        self.apps=apps if apps else dict()
-        self.app_connections = app_connections if app_connections else dict()
-        self.network_endpoints = network_endpoints
-        self.app_start_order = app_start_order
-        self.digraph = None
-
-    def __rich_repr__(self):
-        yield "apps", self.apps
-        yield "app_connections", self.app_connections
-        yield "network_endpoints", self.network_endpoints
-        yield "app_start_order", self.app_start_order
-
-    def get_fragment_producers(self):
-        """Get a list of all the fragment producers in the system"""
-        all_producers = []
-        all_geoids = set()
-        for app in self.apps.values():
-            console.log(app)
-            producers = app.modulegraph.fragment_producers
-            for producer in producers.values():
-                if producer.geoid in all_geoids:
-                    raise ValueError(f"GeoID {producer.geoid} has multiple fragment producers")
-                all_geoids.add(producer.geoid)
-                all_producers.append(producer)
-        return all_producers
-
-    def make_digraph(self):
-        deps = nx.DiGraph()
-
-        for app_name in self.apps.keys():
-            deps.add_node(app_name)
-
-        for from_app_n, from_app in self.apps.items():
-            for from_ep in from_app.modulegraph.endpoints.values():
-                if from_ep.direction == Direction.OUT:
-                    print("OUT", from_app_n, from_ep.external_name)
-                    for to_app_n, to_app in self.apps.items():
-                        for to_ep in to_app.modulegraph.endpoints.values():
-                            if to_ep.direction == Direction.IN:
-                                if from_ep.external_name == to_ep.external_name:
-                                    deps.add_edge(from_app_n, to_app_n, label=to_ep.external_name)
-                elif from_ep.direction == Direction.IN:
-                    print("IN", from_app_n, from_ep.external_name)
-
-
-        return deps
-
-
-    def export(self, filename):
-        self.digraph = self.make_digraph()
-        nx.drawing.nx_pydot.write_dot(self.digraph, filename)
 
 ########################################################################
 #
@@ -542,10 +288,10 @@ def set_mlt_links(the_system, mlt_app_name="trigger", verbose=False):
     if verbose:
         console.log(f"Adding {len(mlt_links)} links to mlt.links: {mlt_links}")
     old_mlt = deepcopy(the_system.apps[mlt_app_name].modulegraph.get_module("mlt"))
-    the_system.apps[mlt_app_name].modulegraph.reset_module("mlt", Module(plugin=old_mlt.plugin,
-                                                                         conf=mlt.ConfParams(links=mlt_links,
-                                                                                             initial_token_count=old_mlt.conf.initial_token_count),
-                                                                         connections=old_mlt.connections))
+    the_system.apps[mlt_app_name].modulegraph.reset_module("mlt", DAQModule(plugin=old_mlt.plugin,
+                                                                            conf=mlt.ConfParams(links=mlt_links,
+                                                                                                initial_token_count=old_mlt.conf.initial_token_count),
+                                                                            connections=old_mlt.connections))
 
 
 def connect_fragment_producers(app_name, the_system, verbose=False):
@@ -676,14 +422,14 @@ def add_network(app_name, the_system, partition_name, verbose=False):
             if verbose:
                 console.log(f"Adding QueueToNetwork named {qton_name} connected to {from_endpoint} in app {app_name}")
 
-            modules_with_network[qton_name] = Module(plugin="QueueToNetwork",
-                                                     connections={}, # No outgoing connections
-                                                     conf=qton.Conf(msg_type=conn.msg_type,
-                                                                    msg_module_name=conn.msg_module_name,
-                                                                    sender_config=nos.Conf(ipm_plugin_type="ZmqPublisher" if type(conn) == Publisher else "ZmqSender",
-                                                                                           address=the_system.network_endpoints[conn_name],
-                                                                                           topic="foo",
-                                                                                           stype="msgpack")))
+            modules_with_network[qton_name] = DAQModule(plugin="QueueToNetwork",
+                                                        connections={}, # No outgoing connections
+                                                        conf=qton.Conf(msg_type=conn.msg_type,
+                                                                       msg_module_name=conn.msg_module_name,
+                                                                       sender_config=nos.Conf(ipm_plugin_type="ZmqPublisher" if type(conn) == Publisher else "ZmqSender",
+                                                                                              address=the_system.network_endpoints[conn_name],
+                                                                                              topic="foo",
+                                                                                              stype="msgpack")))
             # connect the module to the QueueToNetwork
             mod_connections = modules_with_network[from_endpoint_module].connections
             mod_connections[from_endpoint_sink] = Connection(f"{qton_name}.input")
@@ -708,13 +454,13 @@ def add_network(app_name, the_system, partition_name, verbose=False):
                     if verbose:
                         console.log(f"Adding NetworkToQueue named {ntoq_name} connected to {to_endpoint} in app {app_name}")
 
-                    modules_with_network[ntoq_name] = Module(plugin="NetworkToQueue",
-                                                             connections={
-                                                                 "output": Connection(to_endpoint)},
-                                                             conf=ntoq.Conf(msg_type=conn.msg_type,
-                                                                            msg_module_name=conn.msg_module_name,
-                                                                            receiver_config=nor.Conf(name=ntoq_name,
-                                                                                                     subscriptions=["foo"])))
+                    modules_with_network[ntoq_name] = DAQModule(plugin="NetworkToQueue",
+                                                                connections={
+                                                                    "output": Connection(to_endpoint)},
+                                                                conf=ntoq.Conf(msg_type=conn.msg_type,
+                                                                               msg_module_name=conn.msg_module_name,
+                                                                               receiver_config=nor.Conf(name=ntoq_name,
+                                                                                                        subscriptions=["foo"])))
 
         if hasattr(conn, "receiver") and app_name == conn.receiver.split(".")[0]:
             # We're a receiver. Add a NetworkToQueue of receiver type
@@ -730,12 +476,12 @@ def add_network(app_name, the_system, partition_name, verbose=False):
 
             if verbose:
                 console.log(f"Adding NetworkToQueue named {ntoq_name} connected to {to_endpoint} in app {app_name}")
-            modules_with_network[ntoq_name] = Module(plugin="NetworkToQueue",
-                                                     connections={
-                                                         "output": Connection(to_endpoint)},
-                                                     conf=ntoq.Conf(msg_type=conn.msg_type,
-                                                                    msg_module_name=conn.msg_module_name,
-                                                                    receiver_config=nor.Conf(name=ntoq_name)))
+            modules_with_network[ntoq_name] = DAQModule(plugin="NetworkToQueue",
+                                                        connections={
+                                                            "output": Connection(to_endpoint)},
+                                                        conf=ntoq.Conf(msg_type=conn.msg_type,
+                                                                       msg_module_name=conn.msg_module_name,
+                                                                       receiver_config=nor.Conf(name=ntoq_name)))
 
     if unconnected_endpoints:
         # TODO: Use proper logging
